# feat: Use lookup tables for julian date format (3-5% improvement)