This change introduces a lookup table that's built on first access for julian date format used when generating the lineitem table. The optimization eliminates a path that we spend about 14% of total running time in (profile sampled).

Here's a before profile where 14% of the time was spent calling `julian`.
![before-julian-date-optim](https://github.com/user-attachments/assets/9b8a6d9f-f77d-4d80-95a2-2850a06ea18a)

An after profile with the block entirely gone due to lazily initializing it once and using a lookup table for the rest of the dates.
![after-julian-date-optim](https://github.com/user-attachments/assets/d8be3a4f-e6b9-41dc-83d5-f0cb4b9b6211)

The trade-off here is entirely about moving runtime CPU cost to memory which is not a lot (I estimated about 9 KiB from `4 * 2557` where 2557 is the number of `i32`'s we store in the lookup table.

Below are the timings captured from `main` and this branch on an 9950X workstation (single core used)  from a `release` build.

`main` :

```
clflushopt@workstation ~/P/tpchgen-rs (main)> time target/release/tpchgen-cli -s 1 --output-dir=/tmp/tpchdbgen-rs
Generation complete!

________________________________________________________
Executed in    3.74 secs    fish           external
   usr time    3.51 secs  201.00 micros    3.51 secs
   sys time    0.22 secs   97.00 micros    0.22 secs

```

This branch:

```
clflushopt@workstation ~/P/tpchgen-rs (cl/feat/julian-date-lookup-tables)> time target/release/tpchgen-cli -s 1 --output-dir=/tmp/tpchdbgen-rs
Generation complete!

________________________________________________________
Executed in    3.60 secs    fish           external
   usr time    3.39 secs    0.00 micros    3.39 secs
   sys time    0.21 secs  332.00 micros    0.21 secs
```

Edit: Had to replace SVG's with snapped pictures but the profile can be replicated using `cargo flamegraph` & `perf` as such `cargo flamegraph -- -s 1 --output-dir=/tmp/tpchdbgen-rs`